+++ relay/kusama/constants/src/lib.rs
@@ -134,6 +134,97 @@ pub mod system_parachain {
 /// Kusama Treasury pallet instance.
 pub const TREASURY_PALLET_ID: u8 = 18;
 
+pub mod proxy {
+	use pallet_remote_proxy::ProxyDefinition;
+	use polkadot_primitives::{AccountId, BlakeTwo256, BlockNumber, Hash};
+	use sp_runtime::traits::Convert;
+
+	/// The type used to represent the kinds of proxying allowed.
+	#[derive(
+		Copy,
+		Clone,
+		Eq,
+		PartialEq,
+		Ord,
+		PartialOrd,
+		codec::Encode,
+		codec::Decode,
+		codec::MaxEncodedLen,
+		core::fmt::Debug,
+		scale_info::TypeInfo,
+		Default,
+	)]
+	pub enum ProxyType {
+		#[codec(index = 0)]
+		#[default]
+		Any,
+		#[codec(index = 1)]
+		NonTransfer,
+		#[codec(index = 2)]
+		Governance,
+		#[codec(index = 3)]
+		Staking,
+		// Index 4 skipped. Formerly `IdentityJudgement`.
+		#[codec(index = 5)]
+		CancelProxy,
+		#[codec(index = 6)]
+		Auction,
+		#[codec(index = 7)]
+		Society,
+		#[codec(index = 8)]
+		NominationPools,
+		#[codec(index = 9)]
+		Spokesperson,
+		#[codec(index = 10)]
+		ParaRegistration,
+	}
+
+	/// Remote proxy interface that uses the relay chain as remote location.
+	pub struct RemoteProxyInterface<LocalProxyType, ProxyDefinitionConverter>(
+		core::marker::PhantomData<(LocalProxyType, ProxyDefinitionConverter)>,
+	);
+
+	impl<
+			LocalProxyType,
+			ProxyDefinitionConverter: Convert<
+				ProxyDefinition<AccountId, ProxyType, BlockNumber>,
+				Option<ProxyDefinition<AccountId, LocalProxyType, BlockNumber>>,
+			>,
+		> pallet_remote_proxy::RemoteProxyInterface<AccountId, LocalProxyType, BlockNumber>
+		for RemoteProxyInterface<LocalProxyType, ProxyDefinitionConverter>
+	{
+		type RemoteAccountId = AccountId;
+
+		type RemoteProxyType = ProxyType;
+
+		type RemoteBlockNumber = BlockNumber;
+
+		type Hash = Hash;
+
+		type Hasher = BlakeTwo256;
+
+		fn block_to_storage_root(
+			validation_data: &polkadot_primitives::PersistedValidationData,
+		) -> Option<(Self::RemoteBlockNumber, <Self::Hasher as sp_core::Hasher>::Out)> {
+			Some((validation_data.relay_parent_number, validation_data.relay_parent_storage_root))
+		}
+
+		fn local_to_remote_account_id(local: &AccountId) -> Option<Self::RemoteAccountId> {
+			Some(local.clone())
+		}
+
+		fn remote_to_local_proxy_defintion(
+			remote: ProxyDefinition<
+				Self::RemoteAccountId,
+				Self::RemoteProxyType,
+				Self::RemoteBlockNumber,
+			>,
+		) -> Option<ProxyDefinition<AccountId, LocalProxyType, BlockNumber>> {
+			ProxyDefinitionConverter::convert(remote)
+		}
+	}
+}
+
 #[cfg(test)]
 mod tests {
 	use super::{